Revert JACK Futex changes
This reverts commit b6081a11b84c14a0e368ee6bb2f61ec6bc4f018d, reversing
changes made to 6846ee83b1cffffbbc5cd1be3e0f501157646d3e. The JACK futex
patch (pull request #219 on the jackaudio/master branch) was squashed
into one commit, and there were additional changes to it that solved an
issue involving private futexes. This patch reverts the original 2
commits, so that the single squashed commit can be cherry-picked on top
of this to mirror the author of the futex patches changes.